reuse close logic
