+++ interface/components/APIKeyList.tsx
@@ -34,6 +34,8 @@ import {
 import { successToast } from "./Toasts";
 import { KeyIcon } from "@heroicons/react/24/outline";
 
+import { useClickOutsideToast } from './useClickOutsideToast';
+
 export type ApiKeyDisplay = ApiKeys & {
   api_key?: string;
   copied?: boolean;
@@ -47,24 +49,7 @@ const APIKeyList = () => {
   const [apiKeyToUpdate, setApiKeyToUpdate] = useState<string | undefined>();
   const { logout, setUserWarning } = useContext(UserContext);
   const toast = useToast();
-  const toastIdRef = useRef<ToastId | undefined>();
-
-  function closeToast() {
-    if (toastIdRef.current) {
-      toast.close(toastIdRef.current)
-    }
-  }
-
-  useEffect(() => {
-    window.addEventListener("click", (e) => {
-      closeToast()
-    });
-    return () => {
-      window.removeEventListener("click", (e) => {
-        closeToast();
-      })
-    }
-  });
+  const { toastIdRef, openToast } = useClickOutsideToast();
 
   useEffect(() => {
     let keysFetched = false;
@@ -88,7 +73,7 @@ const APIKeyList = () => {
   const handleCreateApiKey = async (key: ApiKeyDisplay) => {
     try {
       setCreateApiKeyModal(false);
-      toastIdRef.current = toast(successToast("API Key created successfully!", toast));
+      openToast("API Key created successfully!");
       setApiKeys([...apiKeys, key]);
     } catch (error: any) {
       const msg =
